prevent unnecessary (un-)boxing (found by FindBugs)
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@5278 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/explicit/ExplicitElement.java
@@ -207,10 +207,10 @@ public class ExplicitElement implements AbstractQueryableElement, FormulaReporti
     ExplicitElement newElement = new ExplicitElement(previousElement);
 
     for(String variableName: constantsMap.keySet())
-      newElement.constantsMap.put(variableName, constantsMap.get(variableName).longValue());
+      newElement.constantsMap.put(variableName, constantsMap.get(variableName));
 
     for(String variableName: referenceCount.keySet())
-      newElement.referenceCount.put(variableName, referenceCount.get(variableName).intValue());
+      newElement.referenceCount.put(variableName, referenceCount.get(variableName));
 
     return newElement;
   }
